post

+++ src/main/java/com/stock/xMarket/model/Order.java
@@ -67,18 +67,22 @@ public class Order implements Serializable{
 	private double frozenAmount;
 	
 	@Column
-	private double state;
+	private int state;
 
 	public Order(){}
 
 	
 
-	
+
+
+
+
+
 
 
 	public Order(long orderId, Stock stock, User user, Time time, int type, Date date, int tradeStraregy,
 			int orderAmount, int exchangeAmount, double orderPrice, double exchangeAveragePrice, int cancelNumber,
-			double frozenAmount, double state) {
+			double frozenAmount, int state) {
 		this.orderId = orderId;
 		this.stock = stock;
 		this.user = user;
@@ -100,7 +104,11 @@ public class Order implements Serializable{
 
 
 
-	public double getState() {
+
+
+
+
+	public int getState() {
 		return state;
 	}
 
@@ -109,9 +117,17 @@ public class Order implements Serializable{
 
 
 
-	public void setState(double state) {
+
+
+
+
+	public void setState(int state) {
 		this.state = state;
 	}
+
+
+
+
